Reorder some tuples to put y first
CONFIG_FOO=y and 'default y' and the like are more common than the n
versions, so test for them first.

Turning the tuples into sets would be even better on Python 3, as it
optimizes sets with constant keys into a LOAD_CONST, but it has a
performance penalty on Python 2.