menuconfig: Only list duplicated choice symbols once
When a Kconfig file defined a named choice and was included multiple
times, the choice symbols were listed multiple times in the menuconfig
as well (due to commit 17d7c1e ("menuconfig: Show all symbols at each
menu location for multi.def. choices")).

That's probably not what you want. Tweak it so that each symbol is only
shown once, with the prompt that was used for it at whatever choice
definition location is entered.

Also change how the choice selection is displayed before the choice is
entered, so that the prompt used for the selected symbol at that
particular location is used. Previously, the prompt at the first
definition location for the symbol was always used.

(Note that all of this is only about how things are displayed. It's not
essential to functionality, in case you're writing your own menuconfig.)