Flag removed symbols as changed in sync_deps()
If a symbol is removed from a Kconfig file, it makes sense to flag it as
changed, so that things that still (probably accidentally) depend on it
get rebuilt.

Saw a patch for the C tools with the same effect floating around, so
might as well add it already. The C tools had other brokeness as well
though...