Save existing configuration to .<filename>.old in write_config()
Add a default-True 'save_old' flag to write_config(). If 'save_old' is
True and an existing configuration file is being overwritten, a copy of
the old configuration file is saved to .<filename>.old (e.g.
.config.old) in the same directory.

Errors are ignored, as the old configuration would usually just be a
nice-to-have, and not essential.

The same functionality could be added for minimal configuration files
and headers, but it's probably most useful for configuration files.