setconfig: Add script
This is a simple script for updating configuration values from the
command line, with (optional) checking that the assigned value matches
the actual symbol value afterwards (which it might not if there are
unsatisfied dependencies).

Sample usage:

  $ setconfig FOO_SUPPORT=y BAR_BITS=8

This is useful for patching the configuration in automated build
systems, in a way that's safer than directly patching configuration
files.