Convert some comments to module docstrings in utilities
This makes running pydoc(3) on the utilities helpful.

Reuse the module docstring for the --help text for utilities that use
argparse.

Also fix some copy-paste errors in the all*config.py descriptions and
clean up the language a bit.

Piggyback removal of an optimization in allmodconfig.py that's pretty
irrelevant now.