menuconfig: Prune _shown_nodes() recursion
When looking for visible children of invisible symbol nodes,
_shown_nodes() was always recursing all the way out the leaf nodes.
That's a ton of redundant work, though it doesn't seem to have lead to
noticeable slowness (it does in the GUI menuconfig).

Stop the search at symbol nodes with node.dep = n (from
'if'/'depends on'). Those can never have visible children, because
node.dep gets propagated to prompts.