menuconfig: Remember last saved/loaded path and improve _conf_changed
Remember the last path that was manually saved/loaded instead of
reverting back to standard_config_filename() (e.g. .config).

Remember the path to the last saved minimal configuration separately as
well.

Also improve the _conf_changed behavior when loading a .config within
the interface. Instead of always treating it as needing to be saved,
check if it's outdated, like for the .config file loaded on startup.

Also make the exit message ("No changes to save", etc.) always include
the target .config file, which is helpful. Previously, only the save
message did.