menuconfig: Fix display issue for unsatisfied-deps selected symbol with children
A symbol with unsatisfied direct dependencies can end up with visible
children in an implicit submenu if it is selected (though that generates
a warning), so the optimization in _shown_nodes() isn't safe, and causes
the child nodes to not be shown outside show-all mode.

Just remove the optimization. Trying things out some more, everything's
plenty fast enough anyway.

Checking the direct dependencies of the parent instead would be safe.