Never prepend '.' to $KCONFIG_CONFIG.old
In retrospect, trying to be "helpful" by saving the old version of a
$KCONFIG_CONFIG that does not start with a '.' as e.g. '.config.old'
instead of 'config.old' is a bad idea, because it means that scripts
can't rely on the backup file simply being called $KCONFIG_CONFIG.old.

I spotted this causing compatibility issues in
https://github.com/automate-lfs/jhalfs/commit/a645174fd43ba4eee84089965df85785878e7aa6.

I had Vim backup files and the like in mind originally, but .config.old
is much more likely to be processed by scripts.

This is a small backwards compatibility break, so the major version will
be increased to 11.