Make 11.0.0 release
Simplifies the filename calculation for .old files, in commit 4fed39d
("Never prepend '.' to $KCONFIG_CONFIG.old"). This makes it reliable to
refer to the backup file as $KCONFIG_CONFIG.old in scripts.

This is a backwards-incompatible change when $KCONFIG_CONFIG is set to a
filename that does not start with a '.', so the major version is
increased. Note that $KCONFIG_CONFIG is usually set to
<some path>/.config, so few people should be affected.