Leave unchanged output files untouched
Before writing a configuration file or header file, compare the old
contents of the file against the new contents. If there's no change,
skip the write, to avoid updating the file modification time.

This might avoid triggering redundant rebuilds depending on how the
build system is set up, and could allow for a simpler setup.