menuconfig: Be silent when changing LC_CTYPE to a UTF-8 locale
Having LC_ALL=C and the like in Makefiles is probably pretty common, to
get consistent collation and messages from commands and stuff. It
currently makes menuconfig.py spit out a warning about changing the
locale to a UTF-8 locale though.

Just remove the message, and silently change the locale. The Python 3.7
PEP 538 code is silent as well.