Make set_value() no-change check work for "n"/"m"/"y" strings
Convert "n"/"m"/"y" to 0/1/2 earlier, so that e.g. sym.set_value("y")
is recognized as a no-op when sym.user_value == 2.

This also helps when loading several configuration files that assign
some of the same symbols, because load_config() calls set_value() with
"n"/"m"/"y".

This could also cut down on the number of times a particular warning is
displayed when loading lots of overlapping configuration files.