Convert standard_kconfig() to argparse for better feedback
Stuff like this is not the pinnacle of helpful design, and hides that
the commands actually have long help texts (that can be viewed with
pydoc):

    $ ./menuconfig.py --help
    [Errno 2] No such file or directory: '--help'

Fix it by converting standard_kconfig() to argparse, and add a
'description' argument to it for the command-specific help text. --help
now shows the same help text shown by pydoc, and some other error
messages are improved as well.

Also fix some copy-paste errors and outdated paths in the help texts for
the all*config commands.