guiconfig: Check for visible nodes before turning off show-all in full-tree mode
If there are no visible nodes in the entire tree, guiconfig
automatically starts in show-all mode. Previously, there was no check
for visible nodes when turning off show-all in full-tree mode in this
case, leading to errors.

Fix it by checking if there are any visible nodes in the top menu when
turning off show-all in full-tree mode. If there aren't, show-all stays
on.