Add environmental variables for autoheader path and preamble.
If no specific filename is passed to 'write_autoconf' then the
environmental variable KCONFIG_AUTOHEADER is used if set, and a default
path otherwise.

Likewise, if no specific preamble for the header file is passed then the
environmental variable KCONFIG_AUTOHEADER_PREAMBLE is used if set, and
a default string otherwise.