Make header strings customizable via the environment
If no header string is specified in write_(min_)config() or
write_autoconf(), use the values of the environment variables
KCONFIG_CONFIG_HEADER and KCONFIG_AUTOHEADER_HEADER, respectively, if
set. KCONFIG_AUTOHEADER_HEADER is consistent with KCONFIG_AUTOHEADER
(the header path), which will be added soon.

Using environment variables avoids having to add separate flags to each
tool that writes configuration files or headers.

Like for $prefix and $CONFIG_, store the values of the environment
variables when the Kconfig instance is created, and expose them via
Kconfig.config_header and Kconfig.header_header. This if flexible and
avoids gotchas when working with multiple Kconfig instances.

Also remove the old default header and make the default no header. Less
advertising, but it felt a bit silly to add workarounds to keep it.

Came up in https://github.com/ulfalizer/Kconfiglib/pull/80.