menuconfig: Check that bright named colors are < curses.COLORS
The bright colors in the range 8-15 are not ANSI and are not guaranteed
to be available.

Previously, the code assumed that all named colors are always available
on color terminals, and skipped the check against curses.COLORS. This
led to this error e.g. with TERM=xterm (as opposed to xterm-256color)
and the aquatic theme, which uses 'brightwhite':

    _curses.error: init_pair() returned ERR

Fix it by checking the number returned for named colors against
curses.COLORS as well.

Came up in https://github.com/espressif/esp-idf/issues/4387.