menuconfig: Ignore MENUCONFIG_STYLE on non-color terminals
Assume that the environment is non-standard/broken on terminals that
don't support color, and that any custom styling applied to the
'monochrome' theme is likely to mess things up (though any colors will
be ignored).

The menuconfig really needs more than two colors to not look ugly
anyway, because of the borderless design.