Add a member function version of load_allconfig()
There was no great reason for this function to be a global function to
begin with. Having it as a member function simplifies callers.

load_allconfig() is a bit special in that it fails with sys.exit(),
which is helpful for tools, but that probably doesn't influence where it
belongs.

Keep the global function for backwards compatibility. A deprecation
warning might be added at some point.