Remove redundant 'node.next = None' assignment for the last node in a file
A node already gets its .next pointer set in _parse_block() if it is
either

 1. followed by another node, or

 2. the last node in a menu, choice, or if.

This also works for nodes from 'source'd files, so there's no need to
special-case the ends of them. Remove the 'node.next = None' assignment
for ends of files.

Instead, special-case just the last node in all files, and set its .next
to None in Kconfig._init().

Unlikely to give a noticeable performance improvement. Just tightens up
the chaining logic a bit.