Use 'python3' instead of 'python' in hashbangs
It's not obligatory for distributions to have a 'python' binary these
days, but 'python3' is likely to be available:
https://www.python.org/dev/peps/pep-0394/#for-python-runtime-distributors.
Change all hashbangs to point to 'python3'.

This will break running

    $ ./script.py

on *nix systems that don't have a 'python3' executable. Changing it to

    $ python script.py

is pretty painless at least.

This change won't break Python 2 when installing via 'pip', because
entry_points creates bounce scripts with their own hashbangs.

The major version will be increased, in case someone looks at the major
version but doesn't install via 'pip'.

Fixes: #89