Return a message from Kconfig.write_autoconf()
Like for Kconfig.write_config() and Kconfig.write_min_config(), return a
string from Kconfig.write_autoconf() with a message saying that the
header got saved, or that there were no changes to it. Can be handy in
tools.

Also make the "no change" message for the various files more specific,
by mentioning what type of file it is (configuration, header, etc.)

Return True/False from Kconfig._write_if_changed() to indicate if the
file was updated. This also allows it to be reused in
Kconfig.write_min_config().